fix: 챌린지 상세 정보 가져오는 방법 JPQL로 변경
