+++ src/main/java/com/whyranoid/walkie/repository/JpaRepositorySample.java
@@ -1,0 +1,7 @@
+package com.whyranoid.walkie.repository;
+
+//@Repository
+//public interface ChallengeRepository extends JpaRepository<ChallengeStatus, Long> {
+//        @Query("select s, c from ChallengeStatus s left join Challenge c on c.challengeId = s.challenge.challengeId where c.challengeId = :challengeId and s.walkie.userId = :userId")
+//        ChallengeStatus findChallengeStatusByChallenge(@Param("challengeId")Long challengeId, @Param("userId")Long userId);
+//}
