+++ src/main/java/com/whyranoid/walkie/service/ChallengeService.java
@@ -3,7 +3,6 @@ package com.whyranoid.walkie.service;
 import com.whyranoid.walkie.domain.Challenge;
 import com.whyranoid.walkie.domain.ChallengeStatus;
 import com.whyranoid.walkie.repository.ChallengeRepository;
-import com.whyranoid.walkie.repository.ChallengeRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,14 +12,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ChallengeService{
 
-    private final ChallengeRepositoryImpl challengeRepositoryImpl;
     private final ChallengeRepository challengeRepository;
 
     public List<Challenge> getChallenges(Long userId, char challengeStatus) {
-        return challengeRepositoryImpl.getChallenges(userId, challengeStatus);
+        return challengeRepository.getChallenges(userId, challengeStatus);
     }
 
-    public ChallengeStatus getChallengeDetail(Long challengeId, Long userId) {
-        return challengeRepository.findChallengeStatusByChallenge(challengeId, userId);
+    public List<ChallengeStatus> getChallengeDetail(Long challengeId, Long userId) {
+        return challengeRepository.getDetailChallenge(challengeId, userId);
     }
 }
